modules/net_tools/ldap: Refactor shared code
