+++ lib/ansible/modules/net_tools/ldap/ldap_attr.py
@@ -149,40 +149,29 @@ modlist:
 
 import traceback
 
+from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils._text import to_native
+from ansible.module_utils.ldap import LdapGeneric, gen_specs, HAS_LDAP
+
 try:
     import ldap
-    import ldap.sasl
-
-    HAS_LDAP = True
 except ImportError:
     HAS_LDAP = False
 
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils._text import to_native
-from ansible.module_utils.ldap import gen_specs
 
-
-class LdapAttr(object):
+class LdapAttr(LdapGeneric):
     def __init__(self, module):
+        LdapGeneric.__init__(self, module)
+
         # Shortcuts
-        self.module = module
-        self.bind_dn = self.module.params['bind_dn']
-        self.bind_pw = self.module.params['bind_pw']
-        self.dn = self.module.params['dn']
         self.name = self.module.params['name']
-        self.server_uri = self.module.params['server_uri']
-        self.start_tls = self.module.params['start_tls']
         self.state = self.module.params['state']
-        self.verify_cert = self.module.params['validate_certs']
 
         # Normalize values
         if isinstance(self.module.params['values'], list):
             self.values = map(str, self.module.params['values'])
         else:
             self.values = [str(self.module.params['values'])]
-
-        # Establish connection
-        self.connection = self._connect_to_ldap()
 
     def add(self):
         values_to_add = filter(self._is_value_absent, self.values)
@@ -239,29 +228,6 @@ class LdapAttr(object):
     def _is_value_absent(self, value):
         """ True if the target attribute doesn't have the given value. """
         return not self._is_value_present(value)
-
-    def _connect_to_ldap(self):
-        if not self.verify_cert:
-            ldap.set_option(ldap.OPT_X_TLS_REQUIRE_CERT, ldap.OPT_X_TLS_NEVER)
-
-        connection = ldap.initialize(self.server_uri)
-
-        if self.start_tls:
-            try:
-                connection.start_tls_s()
-            except ldap.LDAPError as e:
-                self.module.fail_json(msg="Cannot start TLS.", details=to_native(e))
-
-        try:
-            if self.bind_dn is not None:
-                connection.simple_bind_s(self.bind_dn, self.bind_pw)
-            else:
-                connection.sasl_interactive_bind_s('', ldap.sasl.external())
-        except ldap.LDAPError as e:
-            self.module.fail_json(
-                msg="Cannot bind to the server.", details=to_native(e))
-
-        return connection
 
 
 def main():
